Changes to support tool applications in non-mono release
The Tool applications such as the Editor, AssetProcessor and
AssetBundler now can work in release configuration when NOT using the
monolithic-permutation.

There are two main changes.
1. Updated the default value for `sys_PakPriority` CVAR to change its
   file search priority to FileFirst in release configuration.
   Previously this was Pak only
2. Updated the Settings Registry Development Settings Overrides to allow
   merging from the engine, each active gem and the active project `Registry` folders when using the release configuration.

Both this changes combined allows the AssetProcessor and Editor to
locate the files within the project directory and to load necessary
settings registry files required for operation. For example the
AssetProcessor(Batch) needs to load the
`AssetProcessorPlatformConfig.setreg` from the `<engine-root>/Registry`
directory in order bootstrap the AP with enabled platforms, exclude
patterns, the location of the engine's and projects scan folders, etc...

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>