Fix ChangeValidate functions not triggering in Property Editors (#17031)
Both the Reflected Property Editor and Document Property Editor weren't
not triggering ChangeValidate events.

For the Reflected Property Editor, the reason why it wasn't triggering
ChangeValidate events was that the
`PropertyRowWidget::HandleChangeValidateAttribute` function incorrectly
casted the ChangeValidate attribute function to a function that returned
a `bool` type. However all of the Change Validate functions return the
type of `AZ::Outcome<void, AZStd::string>`.

The Document Property Editor issue was more involved.

First the signature of the ChangeValidate functions had the form of
`AZ::Outcome<void, AZStd::string>(C::*)(void*, const Uuid&)`.

Now the way the DomUtils API works now when it comes to storing
non-builtin types such as Uuids, it uses Json Serialization to map a
Uuid to an `AZ::Dom::Value`. Now the Json Serializer for a Uuid writes
it out as just a string.

Now when the DocumentSchema tries to invoke the ChangeValidate attribute
callback, it tries to convert the Dom::Value into the **exact** type of
the function signature.

Now as the Dom::Value is stores a string of a Uuid such as "{EA2C3E90-AFBE-44d4-A90D-FAAF79BAF93D}" it would use the Uuid Json Deserializer to convert the string into an `AZ::Uuid`.

However the DomUtils ValueToType uses special case logic when it tries
to convert the Dom Value back it a C++ type which is a reference. In
this case as the second parameter to ChangeValidate is a `const Uuid&`
it triggers the special case logic.

What that logic does, is instead of trying to use Json Serialization
system to "deserialize" the Uuid string back into an AZ::Uuid it tries
to return a reference to an AZ::Uuid stored in the Dom Value.

Now because Uuid uses Json Serialization, it is not directly stored in
the Dom::Value as a type, so reference for it cannot be retrieved.
Due to this the conversion from the Dom Value to `const Uuid&` fails.

The fix for the issue is to take the `ChangeValidate` function that gets
specified in the `EditContext::Attribute` function and wrap it in a
lambda with the signature of
`AZ::Outcome<void,AZStd::string>(C*, void*, Uuid)` where the Uuid is
accepted by value.

In order to support calling the lambda using the DomInvoke API, the
AttributeInvocable class needs to be able to check if the object it is
storing can be called in a member function like manner.
As the AttributeInvocable can store either a raw function pointer, member function
pointer, member data pointer, or a C++ type with an `operator()`, a
`GetPotentialClassType` function has been added to return
the class type if the AttributeInvocable is storing a member function
pointer or a member data pointer. Or if the AttributeInvocable is
storing a function object with an `operator()` it returns the type of
the first argument to that function if that argument is a C++ class.

fixes #16961

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>