Change material type builder to use JSON ext for "all properties" material
The material type builder saves a default material source file as a convenience for a technical artist to be able to enumerate all of the material properties without parsing the material type file. File was originally saved with the material extension which caused warnings whenever opening the main editor. This change renames the extension to JSON so the file is not recognized as a material source file to stop the warnings.

Resolves https://github.com/o3de/o3de/issues/13677

Signed-off-by: Guthrie Adams <guthadam@amazon.com>