O3DE engine Gem rename support
The following changes have been made to all of the Gems that come with
the O3DE engine source code to make it easier to rename a Gem.

Updated all of the Gems CMakeLists.txt to use the "${gem_name}" variable
when referencing the Gem in CMake Targets.

This allows the name of the CMake targets to update whenever the
"gem_name" in the gem.json is changed.

Updated all of the `AZ_DECLARE_MODULE_CLASS` macro invocations to
reference the `O3DE_GEM_NAME` preprocessor token which is use to
substitute the "gem_name" in the gem.json for the extern "C"
`CreateModuleClass_Gem_*` function which the launcher_generator.cmake
invokes in Monolithic builds to create an AZ::Module derived class instance for each active gem.

resolves #17034

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>