Replace all O3DE engine Gems and Gem Templates Gem variable boilerplate
The repeat boilerplate logic in each Gem has been replaced with a CMake
macro called "o3de_gem_setup" which uses the directory of the calling
CMakeLists.txt file to locate the nearest ancestor gem.json file.

It uses that directory of that gem.json file as the `${gem_path}` value.
It reads the "gem_name" field from that gem.json and sets that as the
`${gem_name}` variable.
It also reads the "version" field from that gem.json and sets that as
the `${gem_version}` variable.

Finally it sets the `${pal_dir}` variable based to the equivalent
`<CMakeLists.txt directory>/Platform/<configured platform>/` directory
using the directory of the calling CMakeLists.txt and the current
platform that CMake is being configured for(Windows, Linux, Android,
MacOS, iOS, or a "restricted platform").

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>