Fixes compile errors and cmake errors for linux release
Note that the change to UiBasics was that other things depend on
builders and tools in a non-tool environment.  Since the UiBasics
builders and tools don't contain any actual code or targets, the
dummy aliases are fine to have in any configuration and simplifies
the use of them from other locations.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>