Fixed memory tracking issue with the `HpAllocator::deallocate` function returning a different value for the amount of bytes deallocated, than were allocated.
The issue is that when the HpAllocator is used for aligned allocation of
>=32 bytes, it could find a block that could fulfill the allocation
request, but wasn't aligned.

If the block need to be adjusted for a size that was >0 but less than
the size of block header(16 bytes) + the size of the free node(40
bytes), than instead of being to split the block into two blocks where
the first block would be a free block that gets added to the free list
and the second block would be the actual aligned block, the code would
try to shift the previous block to be large enough such that the next
block would start aligned on the specified alignent using the
`shift_block` function.

This would result in previous block storing a larger size value.

So when the memory associated with the block was deallocated, it would
report a size larger than the size reported when it was allocated.

This fixes an issue where the Tracking the number of allocated bytes in
the ChildAllocatorSchema and SimpleAllocatorSchema would result in a
non-zero value after all of the allocations have been deleted.

This because the `ChildAllocatorSchema::allocate` function relies on
amount of bytes returned from the actual child allocator `allocate()` call
to be matched when the pointer that was allocated is deallocated via the
`ChildAllocatorSchema::deallocate` function which uses the child
allocator `deallocate()` call.

```
20:28:39  [----------] 1 test from MotionSetFixture

20:28:39  [ RUN      ] MotionSetFixture.MeshLoadTest

20:28:39
   D:\workspace\o3de\Code\Framework\AzCore\AzCore/Memory/ChildAllocatorSchema.h(155):
   error: Child Allocator "EMotionFXAllocator": Total allocated bytes is
   less than zero with a value of -16. Was deallocate() invoked with an
   address that is not associated with this allocator? This should never
   occur

   20:28:39
      D:\workspace\o3de\Code\Framework\AzCore\AzCore/UnitTest/TestTypes.h(76):
      error: Expected equality of these values:

      20:28:39    sizeBeforeTestRan

      20:28:39      Which is: 0

      20:28:39    sizeAfterTestRan

      20:28:39      Which is: 18446744073709551600

      20:28:39  for allocator EMotionFXAllocator with 0 allocation
	 records
```

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>