Fix duplicate FileProcessor::AssessAddedFile call
When the threading gods are angry FileProcessor::AssessAddedFile() can get called on the Main thread and Qt thread from here and from the FileWatcher resulting in a lockup in sqlite3 trying to insert the same file.  This usually repros on the first run when there are lots of assets and a higher likelyhood of encountering the race condition.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>