Reduce the gap between DPE property rows n
Changed the spacing and margins for the DPE to look more like the RPE.
This reduces a large gap between the property rows that added to the screen real estate and required extra scrolling.
This was especially a problem for large property stacks, like material component with complex scenes that have 20, 50, or more materials.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>