Allow customization of DXC executable
Resolves #17148 (DXC should be customizable via Settings Registry)
Which enables working around this bug:
https://github.com/o3de/o3de-extras/issues/625

Developers can now pick a custom version of the DXC executable
via the Settings Registry Key:
`/O3DE/Atom/DxcOverridePath`.

Also, introduced the macro `PRE_HLSL_2021`, along with
an explanation in the following file:
/o3de/Gems/Atom/Feature/Common/Assets/ShaderLib/PRE_HLSL_2021.md
TL;DR: If a game project uses a DXC executable that doesn't conform
to the HLSL 2021 standard, then they should globally define the macro
PRE_HLSL_2021  when compiling shaders.

The O3DE default version of DXC is 1.7.2308 which conforms to HLSL 2021.
The problem is that some shaders get compiled to versions of SPIRV
not supported by Adreno GPU drivers. This commit gives the option
to use a different DXC and avoid this issue.

Removed redundant LinearToSRGB conversion functions.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>