Simplifying thumbnail system asynchronous loading.
Switching to the job system instead of qt futures, watchers, and thread pools.

Removed LoadThread function and made Load virtual. The thumbnailer will call load using a job function. The job function will capture the shared pointer to the thumbnail and keep it alive so it should be safe to fire and forget. Changed some state variables to be atomic.

Replaced counter for number of active loads with set of thumbnail share pointers currently being loaded. This will also make it easier to debug which thumbnails have not completed.

Renamed signals, slots, and other functions used throughout the system that had hard to follow naming.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>