Material component to only use custom previews if properties are overridden
This is an optimization that causes the material component to only render custom previews if properties are overridden on the slot that would display the preview.

If no custom or live preview is rendered then the property asset control thumbnail widget will display the thumbnail for the assigned or default material asset.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>