The Android Project Generator tool. (#17257)
The Android Project Generator tool. (#17257)

This PR adds a new feature to the ProjectManager, by means
of a set of python scripts, that automate the creation
of Keystore files and the generation of an Android Project.

The following picture show the location of the menu option
`Open Android Project Generator...`:

And this is how the APG UI looks like:

The UI was developed with the `tkinter` python module because
it facilitates manual execution of this tool from the terminal
using the python version that ships with O3DE.  The initial iteration
of this UI was developed with PySide2, but it required to run from
the Editor and made it impossible to run from the terminal.

What used to be a runtime exception for `o3de.bat android-configue --set-value` for
Keystore passwords has been changed to an informational message. Otherwise this tool
would be more complicated to implement as it would require the usage of a module
like `pexpect` (Which is not part of the O3DE python distribution) to emulate TTY
input.

A `README.md` file was added to explains how to use the tool.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>