XR Initializes after Asset System
This PR makes sure the XR System is initialized
only after the Asset System is initialized.

This change was required because the new version
of the OpenXRVk Gem requires an asset that defines
all actions supported by the application, this asset
is required during XR Session initialization.

The AZ::RHI::Handle template class is now available
in all script contexts. This is necessary because in
an upcoming PR for o3de-extras some OpenXR APIs are
exposed to the behavior context.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>