Only check enabled platforms for valid assets.
The AssetBundler was looking for the first platform each asset had a platform flag for, however this breaks on non-PC platforms for most engine assets which support building on all platforms, but PC assets will not exist on Mac or Linux platforms and so the asset will not be found.  This change makes the asset lookup only consider enabled platforms.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>