Fixes the ALT key being held down after alt-tabbing
When tabbing out of the editor using alt tab, the editor would
not reset the alt key upon return because the key release event
would never reach it.

To fix this, code was added to the QtEventToAzInputMapper to
reset all key states when the application was deactivated
or reactivated.

However, that code was calling the base InputChannel::UpdateState
function, which is non virtual, and the Input Channel for keyboard
is actually a subclass which has special handling.  It needs to
be cast into its actual concrete type and then ProcessRawInputEvent
needs to be called for it to actually reset the modifier key states.

Because of this, the look camera and other components which queried
the modifier key states were not getting the correct state.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>