CHERRY-PICK from dev to point-release/23103 PR #17464
Fixes the ALT key being held down after alt-tabbing (#17464)

* Fixes the ALT key being held down after alt-tabbing

When tabbing out of the editor using alt tab, the editor would
not reset the alt key upon return because the key release event
would never reach it.

To fix this, code was added to the QtEventToAzInputMapper to
reset all key states when the application was deactivated
or reactivated.

However, that code was calling the base InputChannel::UpdateState
function, which is non virtual, and the Input Channel for keyboard
is actually a subclass which has special handling.  It needs to
be cast into its actual concrete type and then ProcessRawInputEvent
needs to be called for it to actually reset the modifier key states.

Because of this, the look camera and other components which queried
the modifier key states were not getting the correct state.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Added test cases for this

```
LIB: libAzToolsFramework.Tests.so
Loading: libAzToolsFramework.Tests.so
OKAY Library loaded: libAzToolsFramework.Tests.so
OKAY Symbol found: AzRunUnitTests
Note: Google Test filter = *ApplicationStateChange*
[==========] Running 5 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 5 tests from All/ModifierKeyEventFixture
[ RUN      ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_alt_l
[       OK ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_alt_l (16 ms)
[ RUN      ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_shift_l
[       OK ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_shift_l (15 ms)
[ RUN      ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_ctrl_l
[       OK ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_ctrl_l (10 ms)
[ RUN      ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_super_l
[       OK ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_super_l (8 ms)
[ RUN      ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_super_r
[       OK ] All/ModifierKeyEventFixture.ModifierKey_During_ApplicationStateChange_Causes_ModifierKeys_Reset/keyboard_key_modifier_super_r (8 ms)
[----------] 5 tests from All/ModifierKeyEventFixture (57 ms total)

[----------] Global test environment tear-down
[==========] 5 tests from 1 test case ran. (58 ms total)
[  PASSED  ] 5 tests.
OKAY AzRunUnitTests() returned 0
Returning code: 0
```

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>