Improves the startup time of AP on linux.
Note that this could also potentially improve it on mac and windows,
as well as metafile processing in general.

Most of this change is just moving the initialization of the catalog
until the file cache is warmed up and then using the file cache to
fix path cases on case sensitive file systems.

The rest of it is fixing the unit tests to account for this -
the mock for the file system wasn't sufficient to actually use this
new code as it would do things like always answer false for
"is directory".

With this change, all unit tests pass in debug on linux, with no
unexpected assertions and no failures.  In addition, AP starts
and idles very quickly without any long freezes during startup.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>