Fixes for windows
Most of the problems were in oversights in existing tests.
For example the asset catalog tests use a virtual file system
but were using CreateDummyFile which makes real files.

Likewise, the asset catalog was using SystemFile instead of FileIO
which was the mock file IO.  So it would be making mock files in memory
but then checking real files on disk, etc.

Note that the Pass-thru file cache is only used in unit tests, or when
specifically asked for by a command line parameter, which is not used
in any of the test scripts.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>