Improve ap startup on linux (#17526)
* Improves the startup time of AP on linux.

Note that this could also potentially improve it on mac and windows,
as well as metafile processing in general.

Most of this change is just moving the initialization of the catalog
until the file cache is warmed up and then using the file cache to
fix path cases on case sensitive file systems.

The rest of it is fixing the unit tests to account for this -
the mock for the file system wasn't sufficient to actually use this
new code as it would do things like always answer false for
"is directory".

With this change, all unit tests pass in debug on linux, with no
unexpected assertions and no failures.  In addition, AP starts
and idles very quickly without any long freezes during startup.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes a warning I'm not sure I created

During Createjobs,  Asset Processor Manager does this:
```cpp
Q_EMIT CreateJobsDurationChanged(sourceAsset.RelativePath().c_str(), sourceAsset.ScanFolderId());
```

However, scan folder ID of type AZ::s64 which hasn't been registered to Qt
serializer's meta object system, so it does not know how to convert it to
a byte array to emit messages across threads.  This causes warnings
and probably causes that message not to be emitted.

* Fixes for windows

Most of the problems were in oversights in existing tests.
For example the asset catalog tests use a virtual file system
but were using CreateDummyFile which makes real files.

Likewise, the asset catalog was using SystemFile instead of FileIO
which was the mock file IO.  So it would be making mock files in memory
but then checking real files on disk, etc.

Note that the Pass-thru file cache is only used in unit tests, or when
specifically asked for by a command line parameter, which is not used
in any of the test scripts.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>