Fixes a crash and improves detach Prefab
This change allows Prefab containers to be set as "Editor Only" which will
cause their children to be exported into the runtime and CTRL+G mode but
not the container itself.

This bug was found with experimenting with prefab containers and detach.

I also noticed that detach was not symmetrical with creating a prefab,
that is, creating a prefeb creates the container and moves the children
into the container, but the default behavior of detach does not
do the opposite.

I reworked detach to do the opposite, but kept the underlying APIs
and functions the same as before for backward compatibility, so all
existing uses of prefab detach will still be have the same.

the only difference is the user facing menu "detach" now has
"detach" (new behavior) and "Detach and keep container entity" (prior
behavior)

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>