Fixes issue #14879
The issue was caused by the fact that the mouse input device was
checking a variable that was tracking how much time since the last
mouse input event, and considering things to be a new input event
if more than a certain amount of time passed.

However, this function was being called for each axis on the Mouse
(So X, Y, Z separately), but the "Time Since last event" variable
was shared across the axes and thus would cause the X axis (the first)
to act differently from the others.

The change does 2 things - one, it makes a per-axis time, and two,
it only emits a "no movement" event artificially if there were no
events this frame (and does so at the end of the frame).

This brings all the channels back into the same behavior - they
all go into the "Pressed" state initially and if you keep moving the
mouse they remain in the "Updated" state until you stop moving the
mouse, at which point they go into the "released" state, followed
by the "Idle" state.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>