Fixes runtime shader bug on Qualcomm GPUs
For Android, it is necessary to compile the shaders with the macro "PRE_HLSL_2021",
which fixes a runtime error where StandardPBR related shaders fail to create their PSO
(Pipeline State Object) in devices with Qualcomm GPUs like some Samsung phones
or Meta Quest devices.
Long Story: Starting with HLSL 2021, the ternary operator: "cond ? true : false" stopped
compiling for vectors and is limited to only scalar variables. HLSL 2021 introduced
new keywords like "select" which function as the ternary operator for vectors.
The problem is that the generated SPIRV binary for the "select" keyword causes
runtime issues in Qualcomm GPUs.
SOLUTION: The macro "PRE_HLSL_2021" was instroduced in all of the shaders that call "select"
with the purpose of NOT calling "select" and defaulting to the ternary operator for vectors,
which also requires the "-HV 2018" command line option so DXC runs with HLSL version 2018 instead
of version 2021. TODO: Remove this in the future once DXC addresses this problem.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>