Fixes a flaky test and possible actual crash (#17826)
Cherry-pick of PR #17826 to development branch from
point release commit 458541f

It turns out that previewing a FBX or other scene
in the asset browser causes the scene to be loaded
and instantiated (ie, prefab processor runs) in an
async thread.  But that thread sets the name of the container Entity
and causes the entity system to send entity name changes from that
async thread, even though the entity is temp and inactive.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>