Fixes crash on Quest devices
This PR disables timeline semaphores for Quest
devices only. Otherwise the runtime crashes.

This was accomplished by leveraging the fact that
the user must flag the Android project as target for
Meta Quest device. This CMake flag is called
`ANDROID_USE_OCULUS_OPENXR`, and this fix leverages
this CMake variable to define the new C++ macro
named DISABLE_TIMELINE_SEMAPHORES.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>