Fixes crash bugs and improves performance when editing entity properties (#17926) (#17938)
* Fixes crash bugs and improves performance when editing entity properties

* Adds a new API to AzToolsFramework::ToolsApplicationAPI to allow to notify
  the application that a specific property of a specific entity has been changed.
* This is different from the previous API which simply broadcast to everyone
  to refresh every component in every property tree in every GUI in every window
  in the entire editor.
* Added a base call in the Base Editor Entity class which calls this API to clean
  up and de-duplicate code.
* Found a crash that causes the image gradient editor to crash and/or not be able
  to modify an image more than once / not be able to change image properties
  after the first time. Not sure if I caused it with these other changes, but this
  fixes it anyway.

Ultimately, this results in a performance improvement when editing entity properties,
including when using manipulators in the viewport to do so.  In my testing, scrubbing
an entity around in the viewport previously tanked my framerate down to 10fps from 60.

After this change, I can scrub an entity around in the viewport and my framerate stays
at 60fps no matter what I do.  This also affects manipulators and other property editors.

In addition, changing individual properties on components in the inspector happens quickly
and doesn't cause the entire editor to freeze up for seconds nor does it cause it to
scroll around wildly and lose your place.



* Update Code/Framework/AzToolsFramework/AzToolsFramework/API/ToolsApplicationAPI.h




* Update Code/Framework/AzToolsFramework/AzToolsFramework/UI/PropertyEditor/ComponentEditor.cpp




* Updates from PR comments



---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Co-authored-by: Gene Walters <32776221+AMZN-Gene@users.noreply.github.com>