Fixes a refresh issue with sliders in the DPE
Also adds a new function to the DPE to allow for handlers
to clean up any internal state they have before new attributes are
set.  Sliders are the only ones using this new function, but if we
find controls that have internal state that causes problems like this
we can update them to function correctly.

Also adds an assert to check for the case where the widgets being used
by the DPE are sending signals while being recycled and cleared.

The slider was the only one I found that caused that assert, but the
assert is also handled and continues without a crash if it hits, so
it should be safer and less crashy even when compiled in profile mode.

The root cause of this problem is documented in the issue it fixes,
which is https://github.com/o3de/o3de/issues/16947 (Issue #16947)

I did try several other approaches, such as resetting the DPE when
given a reset of "EntireTree or "AttributesAndValues" but too many other
control send an "AttributesAndValues" reset when they are updated, so
clearing the entire tree and recycling all controls caused odd behavior
like sliders not being draggable (since it happens while being dragged).

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>