Adds a check for half16, required for ray tracing.
This adds a check to both DX12 and Vulkan to ensure that the hardware
supports half16.  If it does not, the engine will turn off ray tracing
and log the reason why.

If it finds the necessary support, it will turn on ray tracing and also
log that it has done so.

This change also fixes the print out of errors from the DX12
"ensure success" function.  Previously, this function was printing out
just the first letter of the error message because it was feeding a
wchar string into a printf that expects utf8 input.  This change
explicitly calls the wstring version of the error function and explicitly
converts it to utf8 before feeding it to the utf8 AZ_xxxx function.

Doing so means that localized error messages will print correctly in
environments that support them.  While doing so, I've also eliminated
the use of the macroized functions that change their types based on whether
the UNICODE define is set or not.  This code will function no matter what
the defines are and the types will not change.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>