feat: remove legacy perforce plugin (#17994)
* feat: remove legacy perforce plugin
* address compiling problems PerforceSettings
Signed-off-by: Michael Pollind <mpollind@gmail.com>

Note from Nick_L: This moves the P4 implementation to AzToolsFramework/etc, rather than removes it entirely.  one thing to note is that we never actually *linked* to P4API or P4 libraries or dlls, instead, it uses the CLI version of P4 in a subprocess to do all its ops.