Increases timeout for the multiplayer basic connectivity tests.
Jenkins AR was failing at random, and the logs revaled that 15 attempts
were simply not enough for the server to always start 100% of the time.

I've vastly increased the timeout to about 10.5 minutes (since it is
an increasing backoff.)  This only affects automated testing.

I would rather this test fail because something is definitely,
positively wrong, than it fail because the server is busy or some
other process is happening in the background or something.

Note that this will only increase the time these tests take when
the server actually fails to start for some reason, in the normal
working case, the editor will still connect within a few seconds and
the timeout will not hit.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>