Fixes item (re)selection in the Asset Browser Window
Fixes issue https://github.com/o3de/o3de/issues/18067

Also fixes an issue where if you click on an item in the Asset Browser
window when it was already selected, it would not previously update
the inspector to show the selected item details (it would only respond to
selection change).  I tried updating on focus, but focus happens
for many other reasons too like rolling the mouse wheel or scroll bar.

It also makes it so that when you drop something into the viewport,
the viewport gets focus, letting you hit viewport hotkeys (WASD, etc)
as well as the DELETE key, and the key you press goes to the viewport
you dropped into, instead of attempting to, for example, delete the
asset from the asset browser window that the drag started on.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>