Fixes a crash in test LevelEntityComponentCRUD test (#18164) (#18175)
AutomatedTesting\Gem\PythonTests\editor\EditorScripts\BasicEditorWorkflows_LevelEntityComponentCRUD.py

I really have no idea how this didn't constantly crash before.

tTis script opens a New Level Dialog and before it even manages to fully
appear, it auto-accepts it, which calls the GetTemplatePath function.
This function dereferences (without checking), the pointer given by the
ui to the currently selected item in the list view.

There is no default selection so this would always be nullptr.

I don't know when this changed, but this change makes it so that there
is a preferred default selection, and if the default is missing, it uses
the first item in the list as a final failsafe.

Tested by running the script.  It crashed 100% of the time for me before,
and after, it no longer crashes.  Also, the GUI now nicely selects the
default level by default.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>