Fix locale-dependent data files (#18171)
* Fix for locale-dependent data files being read/written
Fixes a variety of issues related to having a locale set on the machine
that uses the comma as the decimal separator.

Fix for locale-dependent data files being read/written
Fixes a variety of issues related to having a locale set on the machine that uses the comma as the decimal separator.

Two main approaches

    For AssetBuilder, always set, globally in the entire application and thus all builders, the "C" locale, which is the culture invariant locale
    For all other applications, create a PALified scoped thread-local locale setter that can be scoped around the parts where files are read or numbers are converted, then add it around all those sensitive places.
        ObjectStream read and write
        Json didn't need handling, as rapidjson already uses a culture invariant number formatter
        Script Canvas read and write (including LUA generation)
        LUA execution (lua will actually fail to work right if its not in the invariant locale)
        Console variable execution (since it comes from a cfg file and that can be considered portable and invariant)
        Added a new test for objectstream that verifies that the invariant locale is working backwards and forwards.
        Added a JSON test that verifies the same.

result: Fixes issue #11139 and also issue #10987 and also issue #11680 and any similar that are to do with loading and saving.

Testing

Reproduced the issues in the above one before fixing it, they work fine after this fix.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>