Fixes script-only "Quickstart" mode. (#18189) (#18198)
Fixes issue : https://github.com/o3de/o3de/issues/18054

This might have been causing all sorts of other issues in other specific
cases too, but this is the one that brought it to my attention.

Root cause analysis (as best as I can tell)

When generating the installer, the cmake script to do so would
copy assets for gems, and in doing so, would glob all the files and
directories in that candidate gem folder over to the target folder.

In order for such "asset only" gems to function, they need to have a
cmakelists.txt generated which created a Builders target.  So it
would generate one...

To do so, it would check each candidate gem folder, and if it found that
the candidate gem folder DID NOT have a cmakelists.txt generated already
(which happens if it contains targets), it would make one that was
"asset only" and would only get run when the gem.json specified that
it was an "asset" gem.

The problem is that Atom consists of many gems-inside-gems, and it
ended up processing the same directory multiple times.  For example
it would process Atom/RHI/DX12 and notice that this did indeed have
real targets (not an asset only gem) and thus leave it alone, not
generating a cmakelists.txt for it.

But then it would process Atom/RHI, and loop over the files in there
and all the directories in there, and end up processing Atom/RHI/DX12
in the context of Atom/RHI.   Since Atom/RHI had no targets, it would
generate a cmakelists.txt for Atom/RHI/DX12 that was "Asset Only",
overwriting the already generated one that was correct.

To fix this, I made it so that the candidate gem files globbed from a
directory do not include subfolders that themselves were also in the
list of candidate gems (and thus would get their own turn at processing).

Testing performed:

transparent and opaque box testing of the full pipeline
* Creating an installer from the source
* using that installer to create a regular AND script only project
* running the editor for those projects
* running the package script for those projects
* inspecting the packaged projects.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>