Fixes crash on close of Asset AssetEditor
Fixes issue: https://github.com/o3de/o3de/issues/14878

There were actually 2 issues one.

One is a use-after-free, it would attempt to talk to the tab
after the tab was deleted, by deleting the tab prematurely in the
save handler, instead of on close.

The other was the assumption that the source control "RequestEdit"
operation was synchronous.  It is not - it exits immediately after
queueing the event.  This caused the code to leave scope and clean up
(including deleting the variables in lambda captures) before the
operation was complete.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>