Fixes an intermittent crash when working with actors.
Actors in the scene would cause a crash "randomly" when
* duplicated
* spawned new instances
* ctrl-g
* FBX Settings panel accesed and changed

The crash was caused by the mesh feature processor
releasing the last reference to the actors stream buffers
when the previewer asked to make a preview of it (which happens
when you save the FBX settings, or view a thumbnail, or when
the mesh changes on disk, etc.)

Note that the crash would not occur when the action was performed
(so not when you viewed a thumbnail or browsed or saved the settings)
but was primed to occur later by that action hence it seeming
like a random crash later.

The main sequence of events leading to the crash is that the
main actors in the scene using the asset were all pointed at the same
atom model instance, which had an internal refcount.

However, the mesh feature processor would either increase the refcount
or decrease it whenever a model was registered to it.

This means it was possible for the actor to register a reference,
but then the thumbnail previewer get the same model, and register it
and release the reference.  This would cause the refcount to go to 0,
unloading all the buffers the actor needed.

Instead, I change the code to always capture a refcount, then release
it if its not needed.

This causes it to go 0 ---> 1 ---> 0 in the case when the buffers
are really not needed by anyone,

0 ---> 1 ---> 2 ---> 1 in the case where they are needed by someone

this also means buffers that were not needed were not actually being
dropped, so may save memory.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>