Fix asset browser freeze and crash when repeatedly searching. (#18249)
* Fix asset browser freeze and crash when repeatedly searching.

Fixes issue https://github.com/o3de/o3de/issues/18051

This was caused by a number of asynchronous things happening
when searches were being performed.  The main problem was actually
a series of out of order operations where, for example, the search
criteria changed before the view was ready to switch between search
and non-search.

This also has a side effect of reducing the flicker when thumbnails
are updating, and speeds up the search process significantly.

It also fixes an unreported bug where repeatedly searching then clearing
the search would result in the table view not updating and being blank.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Update Code/Framework/AzToolsFramework/AzToolsFramework/AssetBrowser/AssetBrowserFilterModel.cpp

Co-authored-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes one of the automated tests

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Co-authored-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>