Fixed bug that occurs when a device supports both "Fragment Shading Rate" and "Fragment Density Map".
The solution is to give exclusive priority to "Fragment Rate Shading" when
defining the Format Capabilities for "RHI::FormatCapabilities::ShadingRate".
By checking only one of them, we avoid color format selection errors when looking
for color formats compatible with RHI::FormatCapabilities::ShadingRate.
Otherwise the runtime may end up using color formats that are ONLY supported for "Fragment Density Map"
when trying to use "Fragment Shading Rate" feature.

This error was evident when enabling the XR Gem, because the
FoveatedImagePass was using a color format for the Shading Rate
attachment that was only compatible with a Density Map attachment.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>