Fix crash after leaving picking mode
fixes issue https://github.com/o3de/o3de/issues/18365

The hook to register and unregister actions happens once during startup,
and is not meant to have any connection to the lifetime of the objects.

The crash was happening because the EditorTransformComponentSelection
object was being destroyed and recreated every time you enter pick mode,
but its action registration captured its 'this' pointer.

usually, such action registration would be placed inside some kind of long
lived singleton "manager" object, aka a system component, but in this case,
it was put in an ephemeral object.  note that this is still a singleton
object, there is only ever one of it, and it is essentially the "manager",
its just special in that sometimes, it does not exist.

to fix this with as little "splash damage" as possible, I kept most
of the code exactly as-is, without modifying any other classes,
but I made the action registration happen in a static function with
no "this" pointer, which means the compiler could find every illegal
instance.

I added a simple AZ::Environment based singleton pattern in there to get
the current instance when available.  This is also an attempt to keep
the "Blast radius" of this change as small as possible, as this is a
well-used pattern and bypasses any chance of there being weird
dll-related issues.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>